moved the qr-onboarding to be outside the qr scanning navigator
