+++ src/screens/qr-onboarding/TutorialContent.tsx
