+++ src/screens/qr-onboarding/views/ItemView.tsx
