+++ src/screens/qr-onboarding/views/Step1.tsx
