+++ src/screens/qr-onboarding/views/Step2.tsx
