+++ src/screens/qr-onboarding/views/Step3.tsx
