+++ src/screens/qr-onboarding/views/Step4.tsx
